+++ src/vue/Dialog/CreatePassword/FolderField.vue
@@ -10,7 +10,7 @@
 
 <template>
     <div class="password-form-field-wrapper password-form-folder-wrapper">
-        <translate tag="label" for="password-folder" say="Folder" icon="folder" class="area-label" />
+        <translate tag="label" for="password-folder" say="Folder" icon="folder" class="area-label"/>
         <input id="password-folder" type="button" class="area-input" :value="valueLabel" @click="chooseFolder">
     </div>
 </template>
@@ -18,33 +18,40 @@
 <script>
     import AbstractField from "@vue/Dialog/CreatePassword/AbstractField";
     import Translate from "@vc/Translate";
-    import API       from '@js/Helper/api';
+    import API from '@js/Helper/api';
     import FolderManager from "@js/Manager/FolderManager";
 
     export default {
         components: {Translate},
-        extends: AbstractField,
+        extends   : AbstractField,
+        props     : {
+            value: [String, Object]
+        },
         data() {
+            let isObjectValue = typeof this.value !== 'string';
+
             return {
-                valueLabel: this.value
-            }
+                isObjectValue,
+                valueLabel: isObjectValue ? this.value.label:this.value
+            };
         },
         mounted() {
-            this.getFolderName().catch(console.error)
+            this.getFolderName().catch(console.error);
         },
         methods: {
             async getFolderName() {
+                if(this.isObjectValue) return;
                 let folder = await API.showFolder(this.value);
 
                 this.valueLabel = folder.label;
-                this.$emit('folder', folder)
+                this.$emit('folder', folder);
             },
             async chooseFolder() {
                 let result = await FolderManager.selectFolder(this.model);
                 if(result) {
-                    this.model = result.id;
+                    this.model = this.isObjectValue ? result:result.id;
                     this.valueLabel = result.label;
-                    this.$emit('folder', result)
+                    this.$emit('folder', result);
                 }
             }
         }